requirements.txt
flask
pandas
matplotlib
numpy
python-dateutil
plotly